requirements.txt
pyyaml
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0
